# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Documentacion de la api escuela",
    "version": "1.1",
    "title": "Api Escuela",
    "contact": {
      "name": "Smart Devs",
      "url": "https://smartdevs.com.gt",
      "email": "informacion@smartdevs.com.gt"
    }
  },
  "host": "localhost:3700",
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "description": "Copiar 'Bearer ' + JWT token (recibe el token loguiando a url \"/usuarios/login\" ) ingresar el resultado en el campo",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/personas/{Id}": {
      "get": {
        "tags": [
          "AD-Personas"
        ],
        "summary": "buscar persona por Id",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_persona.findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de persona",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opeeracion exitosa",
            "schema": {
              "$ref": "#/definitions/ad_personas"
            }
          },
          "400": {
            "description": "Id Invalido"
          },
          "404": {
            "description": "persona No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "AD-Personas"
        ],
        "summary": "Modificar persona",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_persona.update",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de persona",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "objeto persona actualizado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ad_personas"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Id invalido"
          },
          "404": {
            "description": "persona no encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "AD-Personas"
        ],
        "summary": "Eliminar persona",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_persona.delete",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "El id de persona a ser eliminado",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "id invalido"
          },
          "404": {
            "description": "id no encontrado"
          }
        }
      }
    },
    "/personas/buscar/{:Txt}": {
      "get": {
        "tags": [
          "AD-Personas"
        ],
        "summary": "buscar por nombre de persona",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_persona.findLike",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": ":Txt",
            "in": "path",
            "description": "Busqueda por nombre persona",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "schema": {
              "$ref": "#/definitions/ad_personas"
            }
          },
          "400": {
            "description": "campo invalido"
          },
          "404": {
            "description": "registro No encontrado"
          }
        }
      }
    },
...
    "/usuarios/{Id}": {
      "get": {
        "tags": [
          "AD-Usuarios"
        ],
        "summary": "buscar usuario por Id",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_usuario.findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de usuario",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opeeracion exitosa",
            "schema": {
              "$ref": "#/definitions/ad_usuarios"
            }
          },
          "400": {
            "description": "Id Invalido"
          },
          "404": {
            "description": "usuario No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "AD-Usuarios"
        ],
        "summary": "Modificar usuario",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_usuario.update",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de usuario",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "objeto usuario actualizado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ad_usuarios"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Id invalido"
          },
          "404": {
            "description": "usuario no encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "AD-Usuarios"
        ],
        "summary": "Eliminar usuario",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_usuario.delete",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "El id de usuario a ser eliminado",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "id invalido"
          },
          "404": {
            "description": "id no encontrado"
          }
        }
      }
    },
    "/usuarios/buscar/{:Txt}": {
      "get": {
        "tags": [
          "AD-Usuarios"
        ],
        "summary": "buscar por nombre de usuario",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_usuario.findLike",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": ":Txt",
            "in": "path",
            "description": "Busqueda por nombre de usuario",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "schema": {
              "$ref": "#/definitions/ad_usuarios"
            }
          },
          "400": {
            "description": "campo invalido"
          },
          "404": {
            "description": "registro No encontrado"
          }
        }
      }
    },
...
    "/tipos-personas/{Id}": {
      "get": {
        "tags": [
          "AD-Tipos-Personas"
        ],
        "summary": "buscar tipo_persona por Id",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_tipo_persona.findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de tipo_persona",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opeeracion exitosa",
            "schema": {
              "$ref": "#/definitions/ad_tipos_personas"
            }
          },
          "400": {
            "description": "Id Invalido"
          },
          "404": {
            "description": "tipo_persona No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "AD-Tipos-Personas"
        ],
        "summary": "Modificar tipo_persona",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_tipo_persona.update",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de tipo_persona",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "objeto tipo_persona actualizado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ad_tipos_personas"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Id invalido"
          },
          "404": {
            "description": "tipo_persona no encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "AD-Tipos-Personas"
        ],
        "summary": "Eliminar tipo_persona",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_tipo_persona.delete",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "El id de tipo_persona a ser eliminado",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "id invalido"
          },
          "404": {
            "description": "id no encontrado"
          }
        }
      }
    },
    "/tipos-personas/buscar/{:Txt}": {
      "get": {
        "tags": [
          "AD-Tipos-Personas"
        ],
        "summary": "buscar por nombre de la persona",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_tipo_persona.findLike",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": ":Txt",
            "in": "path",
            "description": "Busqueda por nombre de persona",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "schema": {
              "$ref": "#/definitions/ad_tipos_personas"
            }
          },
          "400": {
            "description": "campo invalido"
          },
          "404": {
            "description": "registro No encontrado"
          }
        }
      }
    },
...
    "/acciones/{Id}": {
      "get": {
        "tags": [
          "SA-Acciones"
        ],
        "summary": "buscar accion por Id",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_accion.findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de accion",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opeeracion exitosa",
            "schema": {
              "$ref": "#/definitions/sa_acciones"
            }
          },
          "400": {
            "description": "Id Invalido"
          },
          "404": {
            "description": "accion No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "SA-Acciones"
        ],
        "summary": "Modificar accion",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_accion.update",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de accion",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "objeto accion actualizado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/sa_acciones"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Id invalido"
          },
          "404": {
            "description": "accion no encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "SA-Acciones"
        ],
        "summary": "Eliminar Accion",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_accion.delete",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "El id de accion a ser eliminado",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "id invalido"
          },
          "404": {
            "description": "id no encontrado"
          }
        }
      }
    },
    "/acciones/buscar/{:Txt}": {
      "get": {
        "tags": [
          "SA-Acciones"
        ],
        "summary": "buscar accion por nombre",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_accion.findLike",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": ":Txt",
            "in": "path",
            "description": "Busqueda por nombre de la accion",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "schema": {
              "$ref": "#/definitions/sa_acciones"
            }
          },
          "400": {
            "description": "Id Invalido"
          },
          "404": {
            "description": "accion No encontrado"
          }
        }
      }
    },
...
    "/bitacoras/{Id}": {
      "get": {
        "tags": [
          "SA-Bitacoras"
        ],
        "summary": "buscar bitacora por Id",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_bitacora.findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de bitacora",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opeeracion exitosa",
            "schema": {
              "$ref": "#/definitions/sa_bitacoras"
            }
          },
          "400": {
            "description": "Id Invalido"
          },
          "404": {
            "description": "bitacora No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "SA-Bitacoras"
        ],
        "summary": "Modificar bitacora",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_bitacora.update",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de bitacora",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "objeto bitacora actualizado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/sa_bitacoras"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Id invalido"
          },
          "404": {
            "description": "bitacora no encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "SA-Bitacoras"
        ],
        "summary": "Eliminar bitacora",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_bitacora.delete",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "El id de bitacora a ser eliminado",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "id invalido"
          },
          "404": {
            "description": "id no encontrado"
          }
        }
      }
    },
    "/bitacoras/buscar/{:Txt}": {
      "get": {
        "tags": [
          "SA-Bitacoras"
        ],
        "summary": "buscar bitacora por nombre de la funcion",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_bitacora.findLike",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": ":Txt",
            "in": "path",
            "description": "Busqueda por nombre de la funcion",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opeeracion exitosa",
            "schema": {
              "$ref": "#/definitions/sa_bitacoras"
            }
          },
          "400": {
            "description": "Campo Invalido"
          },
          "404": {
            "description": "registro no encontrado"
          }
        }
      }
    },
...
    "/detalles-funciones/{Id}": {
      "get": {
        "tags": [
          "SA-Detalles-Funciones"
        ],
        "summary": "buscar detalle_funcion por Id",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_detalle_funcion.findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de detalle_funcion",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opeeracion exitosa",
            "schema": {
              "$ref": "#/definitions/sa_detalles_funciones"
            }
          },
          "400": {
            "description": "Id Invalido"
          },
          "404": {
            "description": "detalle_funcion No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "SA-Detalles-Funciones"
        ],
        "summary": "Modificar detalle_funcion",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_detalle_funcion.update",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de detalle_funcion",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "objeto detalle_funcion actualizado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/sa_detalles_funciones"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Id invalido"
          },
          "404": {
            "description": "detalle_funcion no encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "SA-Detalles-Funciones"
        ],
        "summary": "Eliminar detalle_funcion",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_detalle_funcion.delete",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "El id de detalle_funcion a ser eliminado",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "id invalido"
          },
          "404": {
            "description": "id no encontrado"
          }
        }
      }
    },
    "/detalles-funciones/buscar/{:Txt}": {
      "get": {
        "tags": [
          "SA-Detalles-Funciones"
        ],
        "summary": "buscar detalles funciones por ruta",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_detalle_funcion.findLike",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": ":Txt",
            "in": "path",
            "description": "Busqueda por por ruta en detalles funciones",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "schema": {
              "$ref": "#/definitions/sa_detalles_funciones"
            }
          },
          "400": {
            "description": "Id Invalido"
          },
          "404": {
            "description": "elemento No encontrado"
          }
        }
      }
    },
...
    "/funciones/{Id}": {
      "get": {
        "tags": [
          "SA-Funciones"
        ],
        "summary": "buscar funcion por Id",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_funcion.findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de funcion",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Opeeracion exitosa",
            "schema": {
              "$ref": "#/definitions/sa_funciones"
            }
          },
          "400": {
            "description": "Id Invalido"
          },
          "404": {
            "description": "funcion No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "SA-Funciones"
        ],
        "summary": "Modificar funcion",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_funcion.update",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "Id de funcion",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "objeto funcion actualizado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/sa_funciones"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Id invalido"
          },
          "404": {
            "description": "funcion no encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "SA-Funciones"
        ],
        "summary": "Eliminar funcion",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_funcion.delete",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "El id de funcion a ser eliminado",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "id invalido"
          },
          "404": {
            "description": "id no encontrado"
          }
        }
      }
    },
    "/funciones/buscar/{:Txt}": {
      "get": {
        "tags": [
          "SA-Funciones"
        ],
        "summary": "buscar funciones por nombre",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_funcion.findLike",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": ":Txt",
            "in": "path",
            "description": "Busqueda por nombre en funcion",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "schema": {
              "$ref": "#/definitions/sa_funciones"
            }
          },
          "400": {
            "description": "campo invalido"
          },
          "404": {
            "description": "elemento No encontrado"
          }
        }
      }
    },
...
    "/funciones-acciones/buscar/{:Txt}": {
      "get": {
        "tags": [
          "SA-Funciones-Acciones"
        ],
        "summary": "buscar funciones por nombre de la funcion",
        "description": "Esto solo puede ser realizado por un usuario autenticado y que tenga los privilegios requeridos.",
        "operationId": "c_funcion_accion.findLike",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": ":Txt",
            "in": "path",
            "description": "Busqueda por nombre de la funcion",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "schema": {
              "$ref": "#/definitions/sa_funciones_acciones"
            }
          },
          "400": {
            "description": "campo invalido"
          },
          "404": {
            "description": "registro No encontrado"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ad_personas": {
      "type": "object",
      "properties": {
        "Nombres": {
          "type": "string"
        },
        "Apellidos": {
          "type": "string"
        },
        "Identificacion": {
          "type": "string"
        },
        "Nacimiento": {
          "type": "string",
          "format": "date-time"
        },
        "Estado": {
          "type": "string"
        },
        "TCEmpresaId": {
          "type": "string",
          "maxLength": 36
        }
      }
    },
    "ad_usuarios": {
      "type": "object",
      "properties": {
        "Email": {
          "type": "string"
        },
        "Nombre": {
          "type": "string"
        },
        "Clave": {
          "type": "string"
        },
        "Intentos": {
          "type": "integer",
          "maximum": 11
        },
        "Reinicio": {
          "type": "integer",
          "maximum": 1
        },
        "Estado": {
          "type": "string"
        },
        "TCEmpresaId": {
          "type": "string",
          "maxLength": 36
        },
        "SARolId": {
          "type": "string",
          "maxLength": 36
        },
        "ADTipoPersonaId": {
          "type": "string",
          "maxLength": 36
        }
      }
    },
    "ad_tipos_personas": {
      "type": "object",
      "properties": {
        "Nit": {
          "type": "string"
        },
        "Estado": {
          "type": "string"
        },
        "TCEmpresaId": {
          "type": "string",
          "maxLength": 36
        },
        "TCMunicipioId": {
          "type": "string",
          "maxLength": 36
        },
        "ADPersonaId": {
          "type": "string",
          "maxLength": 36
        }
      }
    },
...
  }
}